Phase 2a: Extract resume subcommand into commands/resume.py
Moves run_resume_command into its own module, following the same
pattern as build.py. The package re-exports the name so callers do
not change.

resume.py only needs argparse, os, sys, pathlib.Path, typing.Optional,
checkpoint helpers, logging_setup helpers, and load_settings — a
significantly tighter import surface than commands/__init__.py provided.

Cleans up imports in __init__.py that are no longer used (Optional,
checkpoint helpers).
